+++ R/convert_to.R
@@ -6,6 +6,7 @@
 #' @param template a string with the template name.
 #' @param parse_fields a logical indicating if the fields must be parsed.
 #' @param format output format
+#' @param destdir a string with destination directory to save converted file
 #'
 #' @return a string with the file path of generated file.
 #'
@@ -19,11 +20,11 @@
 #' }
 #' @export
 convert_to <- function(filename, template = NULL, parse_fields = TRUE,
-                       format = "csv") {
+                       format = "csv", destdir = NULL) {
   template <- .retrieve_template(filename, template)
   fname <- Filename(name = filename)
   df <- template$read_file(filename, parse_fields)
-  new_filename <- fname$changeExt(format)
+  new_filename <- fname$changeExt(paste0(".", format), destdir)
   if (format == "csv") {
     write.table(df,
       file = new_filename, sep = ",", dec = ".",