+++ NEWS.md
@@ -1,6 +1,7 @@
 # rb3 0.0.1.9000
 
 * changes for ropensci process, replaced `sapply` with `purrr::map_xxx`
+* added argument `destdir = NULL` to `convert_to` function.
 
 # rb3 0.0.1